refactor: phase 6 - move api and ai routes to api blueprint
