+++ src/app.py
@@ -26,10 +26,8 @@ from core.reporter import generate_pdf_report
 from core.deep_subdomain_scanner import scan_subdomains_blocking
 from core.database_vulnerability_scanner import scan_database_vulnerabilities_blocking
 from core.directory_scanner import scan_directories_blocking
-from core.mapper import TopologyMapper
 from core.osint_engine import OSINTEngine
 from core.whois_lookup import WhoisLookup
-from services.ai_service import generate_port_analysis, generate_db_analysis
 from services.storage_service import HISTORY_FILE, load_history, save_history
 # Configure Flask app
 app = Flask(__name__, 
@@ -166,29 +164,6 @@ def database_vulnerability_page():
     
     return render_template('database_vulnerability.html', vulnerabilities=vulnerabilities, message=message, active_page='database-vulnerability')
 
-@app.route('/api/save-settings', methods=['POST'])
-def save_settings_api():
-    """API endpoint to save scanner settings to session"""
-    try:
-        settings = request.get_json()
-        # Store settings in session
-        session['scanner_settings'] = settings
-        return jsonify({'status': 'success', 'message': 'Settings saved'}), 200
-    except Exception as e:
-        logger.error(f"Error saving settings: {e}")
-        return jsonify({'status': 'error', 'message': str(e)}), 500
-
-@app.route('/api/get-settings', methods=['GET'])
-def get_settings_api():
-    """API endpoint to retrieve scanner settings from session"""
-    try:
-        settings = session.get('scanner_settings', {})
-        return jsonify(settings), 200
-    except Exception as e:
-        logger.error(f"Error retrieving settings: {e}")
-        return jsonify({'status': 'error', 'message': str(e)}), 500
-
-
 @app.route('/export/<scan_id>', methods=['GET'])
 def export_report(scan_id):
     history = load_history()
@@ -209,45 +184,6 @@ def export_report(scan_id):
 @app.route('/topology')
 def topology_page():
     return render_template('topology.html', active_page='topology')
-
-@app.route('/api/topology-data')
-def api_topology_data():
-    # In a real scenario, this pulls from your scan history
-    # For testing, we return a structured graph
-    mapper = TopologyMapper() 
-    return jsonify(mapper.generate_graph_data())
-
-@app.route('/ai_analysis', methods=['POST'])
-def ai_analysis():
-    """AI-powered security analysis for a specific port using Google Gemini."""
-    data = request.get_json() or {}
-    port = data.get('port')
-    service = data.get('service', 'Unknown')
-    banner = data.get('banner', 'No banner')
-    severity = data.get('severity', 'Low')
-    result = generate_port_analysis(port, service, banner, severity)
-    status_code = result.pop('_status_code', 200)
-    return jsonify(result), status_code
-
-
-@app.route('/db_analysis', methods=['POST'])
-def db_analysis():
-    """AI-powered security analysis for database vulnerabilities using Google Gemini."""
-    data = request.get_json() or {}
-    vuln_name = data.get('name', 'Unknown')
-    vuln_description = data.get('description', '')
-    vuln_evidence = data.get('evidence', '')
-    risk_level = data.get('risk', 'Low')
-    vuln_recommendation = data.get('recommendation', '')
-    result = generate_db_analysis(
-        vuln_name,
-        vuln_description,
-        vuln_evidence,
-        risk_level,
-        vuln_recommendation,
-    )
-    status_code = result.pop('_status_code', 200)
-    return jsonify(result), status_code
 
 @app.route('/download_report', methods=['POST'])
 def download_report():
@@ -946,6 +882,8 @@ def run_db_scan_task(target, deep_scan, custom_threads=None):
 
 from routes.views import views_bp
 app.register_blueprint(views_bp)
+from routes.api import api_bp
+app.register_blueprint(api_bp)
 
 if __name__ == '__main__':
     PORT = int(os.environ.get('PORT', 5000))
